+++ src/aatoolbox/datasources/codab/__init__.py
@@ -1,0 +1,11 @@
+"""Methods pertaining to the COD administrative boundaries.
+
+`Common Operational Datasets <https://cod.unocha.org>`
+(CODs) are definitive reference datasets governed by OCHA Field Information
+Section (FIS) and designed
+to support decision making during a humanitarian response.
+The Administrative Boundary (AB) CODs are geospatial datasets that
+delineate a country's borders and internal regions.
+A key feature of the COD AB datasets are P-codes, which are unique
+alphanumeric identifiers for each geographic region.
+"""
